factor out 'presenters' into separate package to allow different ways to present the data
